pyproject.toml
[project]
name = "py-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index-embeddings-huggingface>=0.7.0",
    "llama-index-llms-huggingface-api>=0.7.0",
]

[tool.uv.workspace]
members = [
    "unit2.1/tests",
]
